Chest TriggerZone Volume created + moved EndZone Voume to Chest location.
The Chest Trigger Zone is to be used to start the Chest animation. This opens the cover animation. Also, the Particle System that is inside it, needs to be off and switched on when the Chest Trigger Zone is triggered by the player to sync with the chest magic.

Created Chest Tag for the Chest Trigger Zone